backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pinecone==5.0.0
sentence-transformers==2.3.1
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.6
google-generativeai==0.3.1
langchain==0.0.350
pillow==10.1.0
requests==2.31.0
numpy==1.26.2
python-dotenv==1.0.0
huggingface-hub==0.20.0

